requirements.txt
ultralytics
fastapi
uvicorn
torch
pydantic
pillow
python-multipart
